cross-platforming static content tests
