+++ server/jvm/src/test/scala/org/http4s/server/staticcontent/WebjarServiceFilterSuite.scala
@@ -21,7 +21,7 @@ import org.http4s.Method.GET
 import org.http4s._
 import org.http4s.syntax.all._
 
-class WebjarServiceFilterSuite extends Http4sSuite with StaticContentShared {
+class WebjarServiceFilterSuite extends Http4sSuite with StaticContentSharedJvm {
   def routes: HttpRoutes[IO] =
     webjarServiceBuilder[IO]
       .withWebjarAssetFilter(webjar =>
